pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ror-matcher"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "xxhash>=3.4",
    "pyyaml>=6.0",
    "click>=8.1",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7",
]

[project.scripts]
ror-matcher = "ror_matcher.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
